Reorganizes project
- Moves the scripts folder into backend and renames to jobs
- Creates a new webapp folder into backend
- New shared folder to share code between jobs and webapp
- Updates Makefile and docker-compose to handle env variable
- Updates CI config
- deps.ts and dev_deps.ts are located at the root of the backend folder
