+++ README.md
@@ -8,6 +8,7 @@ Work in progress
 
 To start the web server:
 ```
+cd backend/webapp
 make run
 ```
 
@@ -16,18 +17,20 @@ make run
 To run lint and test you can respectively launch the command:
 
 ```
+cd backend/webapp
 make lint/test
 ```
 
 ## Background tasks
 
-The "scripts" folder contains scripts that retrieve data from the [football API](https://www.api-football.com/) and persist it into a local database. For the moment we are retrieving some teams and the players of these teams.
+The "jobs" folder contains scripts that retrieve data from the [football API](https://www.api-football.com/) and persist it into a local database. For the moment we are retrieving some teams and the players of these teams.
 We are limited by the number of requests per day. The script allows to start the process in incremental mode and to save the state when the number of requests is reached. We can then restart the script the next day and consider the new data.
 
 ### Running the script
 
-In "scripts" folder: 
+Run the script: 
 ```
+cd backend/jobs
 make run
 ```
 
@@ -36,5 +39,6 @@ make run
 To run lint and test you can respectively launch the command: 
 
 ```
+cd backend/jobs
 make lint/test
 ```
