+++ backend/docker-compose.deps.yml
@@ -4,8 +4,6 @@ services:
   mongodb:
     image: mongo:4.4-bionic
     container_name: football_mongodb
-    env_file:
-      - .env
     environment:
       MONGO_INITDB_ROOT_USERNAME: ${MONGO_USERNAME}
       MONGO_INITDB_ROOT_PASSWORD: ${MONGO_PASSWORD}
