+++ backend/jobs/src/teams/team.ts
@@ -1,5 +1,5 @@
-import { Collection } from "../../deps.ts";
-import { bulkUpsert, executeQuery } from "../utils/mongoUtils.ts";
+import { Collection } from "../../../deps.ts";
+import { bulkUpsert, executeQuery } from "../../../shared/mongoUtils.ts";
 import { FootballApiTeam, Team, Venue } from "./types.ts";
 import { FootballApiResponse } from "../utils/apifootball/types.ts";
 import { fetchData } from "../utils/apifootball/apiFootball.ts";
