+++ backend/jobs/src/utils/apifootball/types.ts
