--- scripts/deps.ts
@@ -1,5 +1,0 @@
-export * as log from "https://deno.land/std@0.93.0/log/mod.ts";
-export { Bson, MongoClient } from "https://deno.land/x/mongo@v0.22.0/mod.ts";
-export { Collection } from "https://deno.land/x/mongo@v0.22.0/src/collection/mod.ts";
-import axiod from "https://raw.githubusercontent.com/roonie007/axiod/master/mod.ts";
-export { axiod };
